Tease apart call and conversation
